Fix compilation on vtiger
for some reason, gcc wants a cast from int64_t to long long while clang does not.

+++ src/server/nautical/tiles/ChartTiles.h
@@ -64,7 +64,7 @@ template <> struct Statistics<Angle<double>> {
   }
 
   void appendBson(mongo::BSONObjBuilder* builder) const {
-    builder->append("count", count);
+    builder->append("count", (long long) count);
     if (count > 0 && vectorSum.norm() > 0.01_kn) {
       builder->append("mean", vectorSum.angle().degrees());
     }
